requirements.txt
# Core API Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI Integration (for LLM calls)
openai>=1.40.0

# Deepgram for Speech-to-Text
deepgram-sdk>=3.0.0

# ElevenLabs for Text-to-Speech
elevenlabs>=1.0.0

# WebSocket support
websockets>=12.0
httpx>=0.25.0

# PDF Processing
pypdf==5.0.1
PyPDF2==3.0.1

# Environment Variables
python-dotenv==1.0.1

# CORS Support (included in fastapi but explicit)
pydantic>=2.9.0
